+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MoE: Added `TrainModuleConfig` ABC
 - Added a `MetricSaverCallback` which just saves metrics at specific intervals to JSON files in the `save_folder`.
 - Added `fixed_steps` option to `Checkpointer` and `Evaluator` callbacks for configuring checkpoints/evals at specific step numbers.
+- Added support for gated attention.
+- Added support for no-global-rope ("GNoPE").
 
 ### Fixed
 
@@ -31,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated stable torch version to 2.9.1, updated versions of underlying libraries in Beaker Images.
 - `olmo_core.io.join_path()` now accepts an arbitrary number of components to join.
 - All `olmo_core.nn` module configs now inherit from a common base class, `ModuleConfig`.
+- Big changes to `olmo_core.model_ladder` API.
 
 ## [v2.4.0](https://github.com/allenai/OLMo-core/releases/tag/v2.4.0) - 2025-11-20